[LT-71] Refactor StripeWebHookController and create a .env file for the secrets
