+++ .env
@@ -1,0 +1,27 @@
+# Database Configuration
+DB_USERNAME=postgres
+DB_PASSWORD=admin@123
+
+# JWT Configuration
+JWT_SECRET=mySecretKey123912738aopsgjnspkmndfsopkvajoirjg94gf2opfng2moknm
+
+# Redis Configuration
+REDIS_HOST=redis-19192.c81.us-east-1-2.ec2.redns.redis-cloud.com
+REDIS_PORT=19192
+REDIS_PASSWORD=nCJR4Ym9V1evwEIpPFfEYUSePE7yIYaB
+
+# Email Configuration
+MAIL_USERNAME=pkmsakiththewmikasl@gmail.com
+MAIL_PASSWORD=ieyrmectamxwoymu
+MAIL_FROM=your_email@gmail.com
+
+# RabbitMQ Configuration
+RABBITMQ_USERNAME=guest
+RABBITMQ_PASSWORD=guest
+
+# MongoDB Configuration
+MONGODB_URI=mongodb+srv://lankatrails_mongo:Admin%40123@lankatrails.mg1hub7.mongodb.net/?retryWrites=true&w=majority&appName=lankatrails
+
+# Stripe Configuration
+STRIPE_SECRET_KEY=sk_test_51S2Xb6DdM2gdqwVSadCRU8EDtNLJdfoIRWqWixiTmoKbLxDjdKLHwusBecV11z9mhZmwffOdjWzFYsVHH08vCPyF00rGj5mCFO
+STRIPE_WEBHOOK_SECRET=whsec_GWcYbkpKss52lkLXLp7pqXnBuObWDjHj
