+++ .gitignore
@@ -59,3 +59,5 @@ build/
 
 #uploads
 /uploads
+
+.env
