+++ pom.xml
@@ -239,6 +239,14 @@
 			<optional>true</optional>
 		</dependency>
 
+		<!-- To read .env files -->
+		<dependency>
+			<groupId>me.paulschwarz</groupId>
+			<artifactId>spring-dotenv</artifactId>
+			<version>4.0.0</version>
+		</dependency>
+
+
 		<!-- Testing -->
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
