+++ src/main/java/com/lankatrails/lankatrails_backend/service/StripeService.java
@@ -1,0 +1,7 @@
+package com.lankatrails.lankatrails_backend.service;
+
+import com.lankatrails.lankatrails_backend.dtos.response.APIResponse;
+
+public interface StripeService {
+    APIResponse<String> handleStripeWebhook(String payload, String sigHeader);
+}
