+++ src/main/resources/application.properties
@@ -19,8 +19,8 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/lankatrails
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.database-platform=org.hibernate.spatial.dialect.postgis.PostgisDialect
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-spring.datasource.username=postgres
-spring.datasource.password=admin@123
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
 
 # JPA/Hibernate Configuration
 spring.jpa.show-sql=true
@@ -32,7 +32,7 @@ spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 spring.jpa.open-in-view=false
 
 # JWT Configuration
-spring.app.jwtSecret=mySecretKey123912738aopsgjnspkmndfsopkvajoirjg94gf2opfng2moknm
+spring.app.jwtSecret=${JWT_SECRET}
 spring.app.jwtExpirationMs=3600000
 spring.app.emailVerificationTokenExpirationMs=900000
 spring.app.jwtCookieName=lankatrails_jwt
@@ -60,9 +60,9 @@ management.endpoint.env.enabled=true
 management.endpoint.beans.enabled=true
 
 # Redis Config
-spring.data.redis.host=redis-19192.c81.us-east-1-2.ec2.redns.redis-cloud.com
-spring.data.redis.port=19192
-spring.data.redis.password=nCJR4Ym9V1evwEIpPFfEYUSePE7yIYaB
+spring.data.redis.host=${REDIS_HOST}
+spring.data.redis.port=${REDIS_PORT}
+spring.data.redis.password=${REDIS_PASSWORD}
 spring.data.redis.ssl.enabled=false
 #spring.cache.redis.key-prefix=lankatrails:
 spring.cache.redis.time-to-live=3600
@@ -74,17 +74,17 @@ spring.servlet.multipart.max-request-size=10MB
 # Gmail SMTP server settings
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
-spring.mail.username=pkmsakiththewmikasl@gmail.com
-spring.mail.password=ieyrmectamxwoymu
+spring.mail.username=${MAIL_USERNAME}
+spring.mail.password=${MAIL_PASSWORD}
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
-spring.mail.from=your_email@gmail.com
+spring.mail.from=${MAIL_FROM}
 
 # RabbitMQ connection
 spring.rabbitmq.host=localhost
 #spring.rabbitmq.port=5672
-spring.rabbitmq.username=guest
-spring.rabbitmq.password=guest
+spring.rabbitmq.username=${RABBITMQ_USERNAME}
+spring.rabbitmq.password=${RABBITMQ_PASSWORD}
 spring.rabbitmq.virtual-host=/
 spring.rabbitmq.stomp-port=61613
 
@@ -101,9 +101,9 @@ spring.websocket.buffer-size-limit=4MB
 
 # MongoDB Configuration
 #spring.data.mongodb.uri=mongodb://localhost:27017/lankatrails
-spring.data.mongodb.uri=mongodb+srv://lankatrails_mongo:Admin%40123@lankatrails.mg1hub7.mongodb.net/?retryWrites=true&w=majority&appName=lankatrails
+spring.data.mongodb.uri=${MONGODB_URI}
 spring.data.mongodb.database=lankatrails
 
 # Stripe Configuration
-stripe.secret-key=sk_test_51S2Xb6DdM2gdqwVSadCRU8EDtNLJdfoIRWqWixiTmoKbLxDjdKLHwusBecV11z9mhZmwffOdjWzFYsVHH08vCPyF00rGj5mCFO
-stripe.webhook-secret=whsec_GWcYbkpKss52lkLXLp7pqXnBuObWDjHj
+stripe.secret-key=${STRIPE_SECRET_KEY}
+stripe.webhook-secret=${STRIPE_WEBHOOK_SECRET}
